# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Inlet API",
    "version": "1.0.0",
    "summary": "The self-hosted feedback collector.",
...
    "license": {
      "name": "MIT"
    }
  },
  "components": {
    "securitySchemes": {
      "projectKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "A project credential: a publishable client key (`ipk_…`) for the feedback flow, or a secret server key (`isk_…`) for everything in its project."
      },
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "inlet_session",
        "description": "The management session cookie set by POST /v1/auth/sign-in."
      }
    },
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "malformed_json",
                  "validation_failed",
                  "unsupported_media_type",
                  "payload_too_large",
                  "unauthenticated",
                  "invalid_credentials",
                  "invalid_api_key",
                  "revoked_api_key",
                  "insufficient_scope",
                  "forbidden",
                  "project_not_found",
                  "feedback_database_not_found",
                  "feedback_database_inaccessible",
                  "submission_not_found",
                  "attachment_not_found",
                  "credential_not_found",
                  "not_found",
                  "name_conflict",
                  "form_not_published",
                  "form_version_unknown",
                  "form_version_mismatch",
                  "form_template_invalid",
                  "stale_draft_revision",
                  "no_previous_version",
                  "intent_not_found",
                  "intent_invalid_token",
                  "intent_expired",
                  "intent_payload_conflict",
                  "submission_deleted",
                  "unknown_question",
                  "missing_required_answer",
                  "invalid_answer",
                  "invalid_option",
                  "invalid_email",
                  "answer_too_long",
                  "client_context_too_large",
                  "unsupported_image_format",
                  "animated_image_rejected",
                  "image_too_many_pixels",
                  "file_too_large",
                  "too_many_uploads",
                  "too_many_attachments",
                  "attachment_reference_invalid",
                  "attachment_already_bound",
                  "upload_failed",
                  "invitation_invalid",
                  "invitation_expired",
                  "invitation_already_redeemed",
                  "last_admin_removal",
                  "rate_limit_exceeded",
                  "internal_error"
                ],
                "description": "Stable machine-readable error code."
              },
              "message": {
                "type": "string",
                "description": "Human-readable explanation."
              },
              "details": {
                "description": "Field- or question-level detail, when relevant.",
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "questionId": {
                      "type": "string"
                    },
                    "path": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            },
            "required": [
              "code",
              "message"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
      "Project": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "creator",
              "viewer"
            ]
          },
          "feedbackDatabaseCount": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "name",
          "role",
          "feedbackDatabaseCount",
          "createdAt",
          "updatedAt"
        ],
        "additionalProperties": false
      },
...
    }
  },
  "paths": {
...
    "/v1/projects": {
      "get": {
        "summary": "List the projects you can access",
        "tags": [
          "Projects"
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Project"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a project",
        "tags": [
          "Projects"
        ],
        "description": "The creating user becomes the project’s first Admin (FR-010).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  }
                },
                "required": [
                  "name"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://192.168.0.19:3000",
      "description": "This deployment"
    }
  ],
  "security": [
    {
      "session": []
    },
    {
      "projectKey": []
    }
  ],
...
}